# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "SCIM (System for Cross-domain Identity Management) 2.0 实现，提供用户和组管理功能",
        "title": "SCIM 2.0 API",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "shoy160",
            "email": "shoy160@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "basePath": "/scim/v2",
    "paths": {
...
        "/scim/v2/CustomResourceTypes": {
            "get": {
                "description": "获取自定义资源类型列表，支持分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CustomResourceTypes"
                ],
                "summary": "列出自定义资源类型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "起始索引",
                        "name": "startIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的自定义资源类型，需要提供ID、名称、端点和模式",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CustomResourceTypes"
                ],
                "summary": "创建自定义资源类型",
                "parameters": [
                    {
                        "description": "自定义资源类型信息",
                        "name": "customResourceType",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CustomResourceType"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.CustomResourceType"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/scim/v2/CustomResourceTypes/{id}": {
            "get": {
                "description": "根据ID获取自定义资源类型的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CustomResourceTypes"
                ],
                "summary": "获取自定义资源类型",
                "parameters": [
                    {
                        "type": "string",
                        "description": "自定义资源类型ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CustomResourceType"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "更新现有的自定义资源类型信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CustomResourceTypes"
                ],
                "summary": "更新自定义资源类型",
                "parameters": [
                    {
                        "type": "string",
                        "description": "自定义资源类型ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "自定义资源类型信息",
                        "name": "customResourceType",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CustomResourceType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CustomResourceType"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据ID删除自定义资源类型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CustomResourceTypes"
                ],
                "summary": "删除自定义资源类型",
                "parameters": [
                    {
                        "type": "string",
                        "description": "自定义资源类型ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/model.CustomResourceType"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/scim/v2/{resourceType}": {
            "get": {
                "description": "获取指定类型的自定义资源列表，支持分页和过滤",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CustomResources"
                ],
                "summary": "列出自定义资源",
                "parameters": [
                    {
                        "type": "string",
                        "description": "资源类型",
                        "name": "resourceType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "起始索引",
                        "name": "startIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "过滤条件",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序顺序",
                        "name": "sortOrder",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ListResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的自定义资源",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CustomResources"
                ],
                "summary": "创建自定义资源",
                "parameters": [
                    {
                        "type": "string",
                        "description": "资源类型",
                        "name": "resourceType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "自定义资源信息",
                        "name": "customResource",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CustomResource"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.CustomResource"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.CustomResource": {
            "description": "自定义资源数据定义",
            "type": "object",
            "properties": {
                "attributes": {
                    "description": "存储自定义资源的属性",
                    "type": "object",
                    "additionalProperties": {}
                },
                "createdAt": {
                    "type": "string"
                },
                "externalId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/model.Meta"
                },
                "resourceType": {
                    "description": "关联到 CustomResourceType 的 ID",
                    "type": "string"
                },
                "schemas": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "model.CustomResourceType": {
            "description": "自定义资源类型定义",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/model.Meta"
                },
                "name": {
                    "type": "string"
                },
                "schema": {
                    "description": "自定义资源的 schema URI",
                    "type": "string"
                },
                "schemaExtensions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SchemaExtension"
                    }
                },
                "schemas": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Detail 错误详情",
                    "type": "string"
                },
                "schemas": {
                    "description": "Schemas SCIM错误Schema",
                    "type": "string"
                },
                "scimType": {
                    "description": "ScimType SCIM错误类型",
                    "type": "string"
                },
                "status": {
                    "description": "Status HTTP状态码",
                    "type": "integer"
                }
            }
        },
...
        "model.ListResponse": {
            "type": "object",
            "properties": {
                "Resources": {
                    "description": "Resources 资源列表"
                },
                "itemsPerPage": {
                    "description": "ItemsPerPage 每页项目数",
                    "type": "integer"
                },
                "schemas": {
                    "description": "Schemas SCIM Schema标识符数组",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "startIndex": {
                    "description": "StartIndex 起始索引（从1开始）",
                    "type": "integer"
                },
                "totalResults": {
                    "description": "TotalResults 总结果数量",
                    "type": "integer"
                }
            }
        },
...
        "model.Meta": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "lastModified": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "resourceType": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "model.SchemaExtension": {
            "type": "object",
            "properties": {
                "required": {
                    "type": "boolean"
                },
                "schema": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "请输入 \"Bearer {token}\" 格式的认证令牌",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}